+++ package.json
@@ -27,7 +27,9 @@
     "react-dom": "18.2.0",
     "react-intersection-observer": "^9.4.2",
     "react-modal-video": "^2.0.0",
+    "react-router-dom": "^6.16.0",
     "react-slick": "^0.29.0",
+    "react-toastify": "^9.1.3",
     "sass": "^1.58.1",
     "slick-carousel": "^1.8.1",
     "wowjs": "^1.1.3",
