+++ public/assets/css/App.css
@@ -25,6 +25,7 @@
   align-items: center;
   justify-content: center;
   font-family: 'Inter', sans-serif;
+  
 }
 
 .AppGlass {
@@ -36,6 +37,7 @@
   gap: 16px;
   grid-template-columns: 11rem auto 20rem;
   overflow: hidden;
+  overflow-y: scroll;
 }
 
 @media screen and (max-width: 1200px) {
